# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Driving School API",
        "description": "Description of Driving School API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/slots/{slotId}/reserve": {
            "post": {
                "tags": [
                    "slots"
                ],
                "summary": "Резервирование слота для авторизованного пользователя",
                "operationId": "6fcccf19b133250d53f9d1a71ad7a83f",
                "parameters": [
                    {
                        "name": "slotId",
                        "in": "path",
                        "description": "ID слота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись на вождение",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrivingSessionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Слот недоступен или с начала занятия прошло более 1 часа!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/SlotNotAvailableException"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SlotStartTimeHasAlreadyPastException"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Невозможно забронировать слот!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserNotActiveException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Слот не найден!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SlotNotFoundException"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/slots/{slotId}/cancel": {
            "put": {
                "tags": [
                    "slots"
                ],
                "summary": "Отмена резервирования слота для авторизованного пользователя",
                "operationId": "cd3c8f8bbf7d26115537032597459eab",
                "parameters": [
                    {
                        "name": "slotId",
                        "in": "path",
                        "description": "ID слота",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Слот",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SlotResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Слот закрыт или с начала занятия прошло более 1 часа!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/SlotAlreadyCompletedException"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SlotStartTimeHasAlreadyPastException"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Невозможно забронировать слот!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserNotActiveException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Слот не найден!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SlotNotFoundException"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SlotStatus": {
                "description": "Статус слота\n        available: Доступно к бронированию,\n        reservation: Забронирован,\n        completed: Закончен,\n        blocked: Заблокирован",
                "type": "string",
                "enum": [
                    "available",
                    "reservation",
                    "completed",
                    "blocked"
                ],
                "x-enum-varnames": [
                    "AVAILABLE",
                    "RESERVATION",
                    "COMPLETED",
                    "BLOCKED"
                ]
            },
            "Transmission": {
                "description": "Тип трансмиссии авто\n        mt: Механическая,\n        at: Автоматическая,\n        amt: Автоматизированная механика (робот),\n        cvt: Вариатор",
                "type": "string",
                "enum": [
                    "mt",
                    "at",
                    "amt",
                    "cvt"
                ],
                "x-enum-varnames": [
                    "MT",
                    "AT",
                    "AMT",
                    "CVT"
                ]
            },
...
            "SlotAlreadyCompletedException": {
                "description": "Ответ на ошибку SlotAlreadyCompletedException",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Слот уже закрыт (занятие проведено)!"
                    }
                },
                "type": "object"
            },
            "SlotNotAvailableException": {
                "description": "Ответ на ошибку SlotNotAvailableException",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Невозможно забронировать слот (возможно, его уже кто-то забронировал)!"
                    }
                },
                "type": "object"
            },
            "SlotNotFoundException": {
                "description": "Ответ на ошибку SlotNotFoundException",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Слот не найден!"
                    }
                },
                "type": "object"
            },
            "SlotStartTimeHasAlreadyPastException": {
                "description": "Ответ на ошибку SlotStartTimeHasAlreadyPastException",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Прошло более 1 часа с начала занятия!"
                    }
                },
                "type": "object"
            },
            "UserNotActiveException": {
                "description": "Ответ на ошибку UserNotActiveException",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Невозможно забронировать слот!"
                    }
                },
                "type": "object"
            },
...
            "CarSlotResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Lada"
                    },
                    "model": {
                        "type": "string",
                        "example": "Kalina"
                    },
                    "color": {
                        "type": "string",
                        "example": "Портвейн"
                    },
                    "state_number": {
                        "type": "string",
                        "example": "А100А099"
                    },
                    "transmission": {
                        "$ref": "#/components/schemas/Transmission"
                    }
                },
                "type": "object"
            },
            "DrivingSessionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "score": {
                        "type": "integer",
                        "example": "5"
                    },
                    "status": {
                        "$ref": "#components/schemas/SlotStatus"
                    },
                    "slot_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "student_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "InstructorSlotResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "bio": {
                        "type": "string",
                        "example": "О себе"
                    },
                    "experience": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "SlotResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01 21:00:00"
                    },
                    "status": {
                        "$ref": "#/components/schemas/SlotStatus"
                    },
                    "instructor": {
                        "$ref": "#/components/schemas/InstructorSlotResource"
                    },
                    "car": {
                        "$ref": "#/components/schemas/CarSlotResource"
                    }
                },
                "type": "object"
            }
        }
    },
...
}